chore: reduce polyfills copy/paste.
