+++ compiler/ir/backend.js/src/org/jetbrains/kotlin/ir/backend/js/Dce.kt
@@ -52,6 +52,11 @@ private fun IrDeclaration.addRootsTo(to: MutableCollection<IrDeclaration>, conte
             getter?.addRootsTo(to, context)
             setter?.addRootsTo(to, context)
         }
+        isEffectivelyExternal() -> {
+            if (!hasJsNativeImplementation()) {
+                to += this
+            }
+        }
         isExported(context) -> {
             to += this
         }
@@ -59,9 +64,6 @@ private fun IrDeclaration.addRootsTo(to: MutableCollection<IrDeclaration>, conte
             // TODO: simplify
             if ((initializer != null && !isKotlinPackage() || correspondingPropertySymbol?.owner?.isExported(context) == true) && !isConstant()) {
                 to += this
-            }
-            if (origin == IrDeclarationOrigin.PROPERTY_BACKING_FIELD) {
-                to += correspondingPropertySymbol!!.owner
             }
         }
         this is IrSimpleFunction -> {
@@ -297,9 +299,16 @@ fun usefulDeclarations(
                     // Skip
                 }
 
+                override fun visitField(declaration: IrField) {
+                    val correspondingPropertySymbol = declaration.correspondingPropertySymbol ?: return super.visitField(declaration)
+                    if (declaration.origin == IrDeclarationOrigin.PROPERTY_BACKING_FIELD && correspondingPropertySymbol.owner.hasJsNativeImplementation()) {
+                        declaration.correspondingPropertySymbol!!.owner.enqueue(declaration, "property backing field")
+                    }
+                    super.visitField(declaration)
+                }
+
                 override fun visitDeclaration(declaration: IrDeclarationBase) {
                     if (declaration !== it) declaration.enqueue(it, "roots' nested declaration")
-
                     super.visitDeclaration(declaration)
                 }
             }
