+++ libraries/stdlib/js/src/kotlin/js.math/acosh.kt
@@ -6,15 +6,13 @@
 @file:JsQualifier("Math")
 package kotlin.js
 
+import kotlin.js.math.defineUpperTaylor2Bound
+
 @PublishedApi
 @JsName("acosh")
 @JsNativeImplementation("""
 if (typeof Math.acosh === "undefined") {
-    var epsilon = 2.220446049250313E-16;
-    var taylor_2_bound = Math.sqrt(epsilon);
-    var taylor_n_bound = Math.sqrt(taylor_2_bound);
-    var upper_taylor_2_bound = 1/taylor_2_bound;
-    
+    $defineUpperTaylor2Bound
     Math.acosh = function(x) {
         if (x < 1)
         {
